+++ rest/src/main/java/com/asterlink/rest/service/SentinelDeviceService.java
@@ -9,12 +9,12 @@ import java.util.Map;
  */
 
 public interface SentinelDeviceService {
-    public int createSentinelDevice(SentinelDevice device);
-    public int createSentinelDeviceBatch(List<SentinelDevice> devices);
+    public long createSentinelDevice(SentinelDevice device);
+    public long createSentinelDeviceBatch(List<SentinelDevice> devices);
     public boolean updateSentinelDevice(Map<String, Object> updates);
-    public String deleteSentinelDevice(int id);
-    public SentinelDevice getSentinelDevice(int id);
+    public String deleteSentinelDevice(long id);
+    public SentinelDevice getSentinelDevice(long id);
     public List<SentinelDevice> getAllSentinelDevices();
-    public String claimSentinelDevice(int deviceId, String password, int clientId);
+    public String claimSentinelDevice(long deviceId, String password, int clientId);
     public List<SentinelDevice> findByClientId(int clientId);
 }
